calculate folder validation_loss_calculation and move related files
