+++ tests/test_CalculateScoresBetweenAllIonmodes.py
@@ -4,7 +4,7 @@ import numpy as np
 import pandas as pd
 
 from ms2deepscore.benchmarking.CalculateScoresBetweenAllIonmodes import CalculateScoresBetweenAllIonmodes
-from ms2deepscore.benchmarking.PredictionsAndTanimotoScores import PredictionsAndTanimotoScores
+from ms2deepscore.validation_loss_calculation.PredictionsAndTanimotoScores import PredictionsAndTanimotoScores
 from tests.test_user_worfklow import load_processed_spectrums, TEST_RESOURCES_PATH
 import random
